pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "simple-rpc"
version = "0.1.0"
description = "Simple framework to create your own gRPC server with grpcio"
authors = [
  {name = "NF-coder", email = "reshes@rambler.ru"}
]
license = "MIT"
readme = "README.md"
packages = [
  { include = "SimpleRPC" },
]
requires-python = ">=3.9"
dependencies = [
  "grpcio == 1.71.0",
  "pydantic >= 2.8.0",
]
